Fixed bugs Removed debug information

+++ software/AIWM_Control/core.h
@@ -54,7 +54,7 @@ public slots:
 protected:
 	Swlp m_swlp;
 	QThread m_swlpThread;
-	uint32_t m_commandForSend;
+	uint8_t m_commandForSend;
 };
 
 #endif // CORE_H
